+++ backend/hqlib/__init__.py
@@ -15,4 +15,4 @@ limitations under the License.
 """
 
 NAME = "HQ"
-VERSION = "2.46.6"
+VERSION = "2.46.7"